GUI Microservice/requirements.txt
Flask
requests
azure-iot-device